restructure of dir files maybe?
