+++ src/games/checkers.py
