+++ src/games/solitare.py
